requirements.txt
Django==4.2.25
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
whitenoise==6.6.0
dj-database-url==2.1.0
google-generativeai==0.8.6
gunicorn==22.0.0
